All functional tests are complete & made-to-work, including manual. A nasty bug for Windows callback interface fixed. Improved error detection/reporting (three new error codes). A couple of smaller bugs fixed. Most warnings of Windows fixed.

+++ lib/sockets/pbpal_sockets.c
@@ -236,7 +236,7 @@ bool pbpal_read_over(pubnub_t *pb)
     pb->left -= pb->readlen;
     pb->readlen = 0;
 
-    if (pbpal_read_len(pb) >= pb->len) {
+    if (pbpal_read_len(pb) >= (int)pb->len) {
         /* If we have read all that was requested, we're done. */
         DEBUG_PRINTF("Read all that was to be read.\n");
         pb->sock_state = STATE_NONE;
